feat(items): render fetched items in table
