+++ lib/constants/constants.ts
@@ -1,4 +1,3 @@
-
 const USER_STORAGE_KEY = 'iota-user'
 const API_BASE_URL = "http://localhost:1337/api";
 const FORM_PARENT_STYLES = {
@@ -7,7 +6,6 @@ const FORM_PARENT_STYLES = {
     backgroundColor: "white",
     overflow: "hidden",
 };
-
 export default {
     USER_STORAGE_KEY,
     API_BASE_URL,
